+++ tests/test_correlator.py
@@ -13,8 +13,8 @@ from yaml import Loader, load
 
 from janus_core.calculations.md import NVE
 from janus_core.calculations.single_point import SinglePoint
-from janus_core.helpers.correlator import Correlator
-from janus_core.helpers.observables import Stress
+from janus_core.processing.correlator import Correlator
+from janus_core.processing.observables import Stress
 
 DATA_PATH = Path(__file__).parent / "data"
 MODEL_PATH = Path(__file__).parent / "models" / "mace_mp_small.model"
