+++ tests/test_utils.py
@@ -8,7 +8,8 @@ import pytest
 
 from janus_core.cli.utils import dict_paths_to_strs, dict_remove_hyphens
 from janus_core.helpers.mlip_calculators import choose_calculator
-from janus_core.helpers.utils import none_to_dict, output_structs
+from janus_core.helpers.struct_io import output_structs
+from janus_core.helpers.utils import none_to_dict
 
 DATA_PATH = Path(__file__).parent / "data/NaCl.cif"
 MODEL_PATH = Path(__file__).parent / "models/mace_mp_small.model"
